Add LastSeen functionality
And fixed up the LogFormatter

+++ src/main/java/dev/forbit/server/interfaces/ClientInterface.java
@@ -15,5 +15,5 @@ public interface ClientInterface {
     /**
      * The System.currentTimeMillis() we last heard from the client
      */
-    int getLastSeen();
+    long getLastSeen();
 }
